fix(recorder): review fixes - small refactor of recorder module
